hw/arm/nxp-s32g: Increase gicv3 redist size
Initially, I had set the gic redist size to 0x80000.
Each cpu needs 2 frames of 0x10000 each, so 0x80000
is only enough to support 4 cpus. If the redist size
is too small, then during secondary CPU initialization
we try to access an unmapped address and cause a
data abort to be generated.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>