hw/char/fsl-linflex: Fix uart interrupt bug
The fsl-linflex serial device can sometimes miss raising an
interrupt, resulting in delayed read from vxworks until the
next interrupt (when another key is pressed). This results
in the device seemingly always being 1 keystroke behind.

This change updates the irq status on all writes to the
device, instead of only for specific registers.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>