hw/arm/wr-arm: add unimp region
Add UNIMP device spanning entire address space of s32g machine.

The important thing to note with this change is writes to the
UNIMP device does not generate a data abort, which may result
in some behavioral differences when compared to running VxWorks
on a real system, since the VxWorks data abort handler causes
VxWorks to reboot, which does not happen after this change.

The benefit of this change is that we can track accesses to the
UNIMP region. I can think of some use cases aligned with the
device access tracking plugin, but have not fully thought
through the solution.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>