xlnx-versal: modify some SPI interrupt mappings
The IRQ mapping of some of the shared peripheral interrupts
in the Versal SoC implementation conflict with the IRQ
mappings found in the Versal SoC Technical Reference Manual
AM011 version 1.6, in table "System Interrupts Table". This
change aligns the IRQ mapping of all devices to the mapping
found in the reference manual. This is done in preparation
to add the mmu-500 SMMU, whose IRQ mapping conflicts with
the current EFUSE_IRQ mapping.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>