versal pmc int: remove unused registers
Due to the address space gaps in the register definitions
in the PMC interconnect CSR model implementation, it is not
space efficient to model the registers as a single array
starting from offset 0. We end up statically allocating
multiple MBs worth of unused memory to fill the gaps between
the sparsely laid out registers. Since HVP does not access
the higher offset registers, this change comments out those
registers to save some space in the PMC_INT_REGS struct.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>