versal: use OBJECT_DECLARE_SIMPLE_TYPE
As part of moving the device implementations for LPD INT CSR,
LPD SLCR, and PMC INT CSR from Xilinx QEMU to stable 9.1, some
of the device models need to be migrated to use the modern QOM
templates for device structures. This eliminates the use of
OBJECT_CHECK in favor of the OBJECT_DECLARE_TYPE macros.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>