cadence_gem: add stream-id property to device
The Cadence GEM implementation does not specify any memory
transaction attributes when performing DMA scatter/gather
operations. This is acceptable in non-virtualized scenarios,
but when the DMA needs to go through IOMMU for translation,
the IOMMU needs to know how to map the device to the memory
context in which it should perform the translation.

This change adds a stream-id property to the GEM device
implementation, which is used to set the requester_id field
of the memory transaction attributes specified by the device
when performing DMA operations. The IOMMU implementation
parses the requester_id field as the stream ID which it
uses to map to the page table root which identifies the
memory context.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>