genet: set REMOTE TX OK and LOCAL RX OK at link up
The REMRXOK and LOCALRXOK bits were not set at reset or
link up time, and appear to never be set by the emulation.

I think the real device does not set these bits immediately
at link up, and the emulation was written to match this
behavior exactly. But I believe the hardware may set these
bits some period of time after link up is reported, but
the device emulation does not do this.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>