callstack plugin: vxworks task introspection
This is the second step toward vxworks os introspection.
Parse the address of vxKernelVars from the symbol table and use its
address to read the taskIdCurrent each time it is modified in the
per-cpu WIND_VARS structure. This is used to accurately match a
given stack modification operation (func call or ret) to the thread
stack which it belongs.