callstack plugin: match LR with expected return addr
Store the return address at function callsite as PC+4.
When a RET instruction is detected, compare the contents
of the link register with the stored (expected) return
address of the stack frame to be popped. If not a match,
look backwards in the stack until a matching record is
found.

The reason for this change is that some function calls do
not have a matching return, for example calls to windExit.

During context switch the scheduler saves the task state,
saving the Program Counter as the current address contained
in the Link Register. When the task resumes execution, it
begins executing at the instruction just after the call to
windExit, without ever executing a RET from the call to
windExit.