callstack plugin: use g ptr array for vcpu cache
Use GPtrArray instead of GArray for vcpu caches.

This fixes some confusion I had with how to initialize GArray that
resulted in smp > 1 being broken. The issue was that I had failed
to initialize the VCPUCache->lock for cpuID > 1. So attempting to
access the VCPUCache structure on CPUs > 1 resulted in never being
able to acquire the lock.